2311601633-tugas-mandiri-python-5/requirements.txt
cycler==0.11.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
scipy==1.8.0
six==1.16.0
